requirements.txt
starlette>=0.37.2
pydantic>=2
rich>=13
numpy>=1
setuptools>=68
requests>=2
properscoring>=0.1
pandas>=2.2.3
bittensor==10.0.1
bittensor-cli==9.17.0
async-substrate-interface==1.5.15
substrate-interface==1.7.11
alembic>=1.14.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36
wandb>=0.19.4
tenacity==9.0.0
google-cloud-logging==3.12.1
uvloop==0.21.0
httpx==0.28.1
httpx[http2]
anyio==4.12.1
numba==0.63.1
PyYAML>=6.0.2
duckdb>=1.1.3

# === Simulation & Data (ported from sn50) ===
pymysql>=1.1.0
statsmodels
scipy
arch
tqdm
coinmetrics-api-client
yfinance
matplotlib
